pyproject.toml
[project]
name = "proyectos-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.124.2",
    "openai>=2.11.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
